requirements.txt
asgiref==3.7.2
boto3==1.34.111
botocore==1.34.111
dj-database-url==2.1.0
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.3.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
psycopg2==2.9.9
pycountry==23.12.11
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2023.3.post1
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.1
wheel==0.43.0
XlsxWriter==3.1.9